base cocktails and base ingredients created, time for serious business

+++ src/main/java/org/cocktailmolotov/partyplanner/dto/Cocktails.java
@@ -6,24 +6,24 @@ import java.util.Map.Entry;
 
 public class Cocktails {
 
-	private HashMap<String,HashMap<String,Integer>> cocktails;
+	private HashMap<String,HashMap<String,Double>> cocktails;
 
 	public Cocktails() {
 		// no attributes needed
 	}
 
-	public HashMap<String,HashMap<String,Integer>> getCocktails() {
+	public HashMap<String,HashMap<String,Double>> getCocktails() {
 		return cocktails;
 	}
 
-	public void setCocktails(HashMap<String,HashMap<String,Integer>> cocktails) {
+	public void setCocktails(HashMap<String,HashMap<String,Double>> cocktails) {
 		this.cocktails = cocktails;
 	}
 
 	@Override
 	public String toString() {
 		String result = new String();
-		for(Entry<String,HashMap<String,Integer>> cocktail : cocktails.entrySet()){
+		for(Entry<String,HashMap<String,Double>> cocktail : cocktails.entrySet()){
 			result += cocktail.toString() + "\n";
 		}
 		return result;
